+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2019-05-14
+
+## Bugfix
+- `SoftReferenceInstanceNameTableColumnGenerator` renders null references correctly instead of throwing an error
+
 ## [0.5.0] - 2019-02-23
 
 ## Added